refactor(hitl): pin canonicalisation and approval TTL move to mur-common
